feat: added small nearley.js grammar and parser with html output
These additions are just experimental, and I've actually tried three
different approaches. The first was antlr, but I struggled to get the
antlr runtime to work in javascript. Antlr is the best possible option
as there exists antlr grammars for all major languages, such that I
wouldn't need to write any grammars myself. Unfortunately, the process
for getting antlr to work in javascript is poorly document and poorly
designed. The files for it online use both require and import
statements, and I'd rather not spend my time figuring out how to
transpile this code with babel/webpack. My second choice was pegjs, for
which there at least exists a javascript grammar, but the grammar
doesn't include whitespace, and so I would need to rewrite it anyways. I
then tried to write my own parser for nearley.js, which appears to be a
the best option, assuming that writing the grammar myself is already
necessary. This works, but I really want to revisit getting antlr to
work, so that I at least have a well designed, full grammar to start
with. Pegjs has a grammar but just as an example, so I'm going with
antlr. I am, however, going to try switching to clojurescript. It has
antlr integration (I believe) and can transpile to javascript. Plus,
I've already learned from my SpacedMath project that Clojure is just
better at dealing with ASTs than JS.