feat: settled on using clojure instaparse with bespoke grammar
I tried clojure antlr, which worked at first, but the grammar was very
far away from what it needed to be in terms of parsing whitespace, and
became simpler to build a grammar from scratch. I plan to implement
features as I run into them, and eventually I should have the features
that I need. I left some of the experimental parsers in the project, but
I'll remove them in the next commit.