feat: implemented html generation for feature set
I need to improve the dilineation of different tokens, like the
different types of keywords. At the moment, all keywords look the same,
which I don't think is great. I also would like for identifiers to
change appearance based on whether or not they're being declared or
referenced. And I would like Component and tag references to be
distinct, which would need to be based on capitalization. I also see
that a lot of highlighting colors the brackets and text with slightly
different off-white colors, so I could look into that, but it would be a
lot of work.